+++ lib/completion.ts
@@ -3,6 +3,20 @@ import { Point, Range } from "atom";
 import * as settings from "./settings";
 import * as data from "./data";
 
+interface SuggestionRequest {
+  editor: AtomCore.TextEditor
+  bufferPosition: TextBuffer.Point
+  scopeDescriptor: AtomCore.ScopeDescriptor
+  prefix: string
+  activatedManually: boolean
+}
+
+interface SuggestionInserted {
+  editor: AtomCore.TextEditor
+  triggerPosition: TextBuffer.IPoint
+  suggestion: Completion.TextSuggestion|Completion.SnippetSuggestion
+}
+
 export function activate() {}
 export function deactivate() {}
 
@@ -132,7 +146,7 @@ function pruneSuggestions(prefix: string, suggestions:
 }
 
 /** A callback which we provide to the autocompletion engine for Atom. */
-async function getSuggestions(args: Completion.Params.SuggestionRequest):
+async function getSuggestions(args: SuggestionRequest):
     Promise<Completion.Suggestions> {
   if(!settings.config.generalSettings.enableCompletion.get()) {
     return [];
@@ -259,7 +273,7 @@ function removeRarityPlaceholder(editor: AtomCore.TextEditor, startPosition: Poi
 }
 
 /** Performs the buffer manipulations necessary following a suggestion insertion. */
-function insertedSuggestion(params: Completion.Params.SuggestionInserted) {
+function insertedSuggestion(params: SuggestionInserted) {
   // Whenever the user opens with quotation marks and accepts a suggestion,
   // two closing quotation marks will be left at the end:
   //    BaseType "Cha" -> accepts "Chaos Orb"