+++ lib/filter-data.ts
@@ -1,8 +1,14 @@
-import { Range, CompositeDisposable, Emitter } from "atom";
+import { Point, Range, CompositeDisposable, Emitter } from "atom";
 import * as path from "path";
 import * as assert from "assert";
 
 import * as data from "./data";
+import * as parser from "./parser";
+
+interface BufferChanges {
+  oldRange: TextBuffer.Range
+  newRange: TextBuffer.Range
+}
 
 /** Handles subscriptions for every buffer, while also managing the data for
  *  item filters. */
@@ -10,9 +16,9 @@ class FilterManager {
   private readonly editor: AtomCore.TextEditor;
   private subscriptions: CompositeDisposable;
   private filterSubs: CompositeDisposable;
-  private changes?: Filter.BufferChanges;
+  private changes?: BufferChanges;
 
-  filter?: Promise<Filter.ItemFilter>;
+  filter?: Promise<Filter.Line[]>;
 
   constructor(editor: AtomCore.TextEditor) {
     this.editor = editor;
@@ -32,6 +38,7 @@ class FilterManager {
     }));
 
     this.subscriptions.add(data.emitter.on("poe-did-update-item-data", () => {
+      console.log("Did get the message.");
       this.processIfFilter();
     }));
 
@@ -80,11 +87,11 @@ class FilterManager {
 
   /** Ensures that the buffer contains a filter prior to processing it.. */
   public processIfFilter() {
-    if(this.isFilter()) this.processFilter;
+    if(this.isFilter()) this.processFilter();
   }
 
   /** Processes the entire item filter from scratch. */
-  private async processFilter() {
+  private processFilter() {
     const oldRange = new Range([0, 0], [0, 0]);
 
     const lastRow = this.editor.getLastBufferRow();
@@ -92,26 +99,107 @@ class FilterManager {
     const lastColumn = lastRowText.length - 1;
     const newRange = new Range([0, 0], [lastRow, lastColumn]);
 
-    this.filter = this.getLineInfo({ oldRange: oldRange, newRange: newRange });
-
-    const lines = await this.filter;
-    emitter.emit<Filter.Params.DataUpdate>("poe-did-process-filter",
-        { editorID: this.editor.buffer.id, lines: lines });
+    this.filter = this.parseLineInfo(this.filter, { oldRange: oldRange,
+        newRange: newRange }, true);
   }
 
   /** Processes only the recent changes to the item filter. */
-  private async processFilterChanges() {
+  private processFilterChanges() {
     if(!this.changes || !this.filter) return;
 
-    this.filter = this.getLineInfo(this.changes);
-    const lines = await this.filter;
-    emitter.emit<Filter.Params.DataUpdate>("poe-did-process-filter",
-        { editorID: this.editor.buffer.id, lines: lines });
+    this.filter = this.parseLineInfo(this.filter, this.changes);
+    this.changes = undefined;
   }
 
-  private async getLineInfo(change: Filter.BufferChanges) {
-    const linterData = await data.linterData;
-    return [];
+  private translateLineRanges(line: Filter.Line, delta: Point) {
+    switch(line.type) {
+      case "Block": {
+        const fb: Filter.Block = (<Filter.Block>line.data);
+        fb.scope = fb.scope.translate(delta);
+        fb.type.range = fb.type.range.translate(delta);
+        if(fb.trailingComment) {
+          fb.trailingComment.range = fb.trailingComment.range.translate(delta);
+        }
+        line.data = fb;
+      } break;
+      case "Comment": {
+        const fb: Filter.Comment = (<Filter.Comment>line.data);
+        fb.range = fb.range.translate(delta);
+      } break;
+      case "Rule": {
+        const fb: Filter.Rule = (<Filter.Rule>line.data);
+        fb.range = fb.range.translate(delta);
+        fb.type.range = fb.type.range.translate(delta);
+        if(fb.operator) fb.operator.range = fb.operator.range.translate(delta);
+        fb.values.forEach((value) => {
+          value.range = value.range.translate(delta);
+        })
+        if(fb.trailingComment) {
+          fb.trailingComment.range = fb.trailingComment.range.translate(delta);
+        }
+      } break;
+      case "Unknown": {
+        const fb: Filter.Unknown = (<Filter.Unknown>line.data);
+        fb.range = fb.range.translate(delta);
+      } break;
+      default:
+        break;
+    }
+  }
+
+  private async parseLineInfo(filter: Promise<Filter.Line[]>|undefined,
+      changes: BufferChanges, reset = false): Promise<Filter.Line[]> {
+    const lines = this.editor.buffer.getLines();
+    const itemData = await data.filterItemData;
+
+    var previousLines: Filter.Line[];
+    if(reset) previousLines = [];
+    else if(filter) previousLines = await filter;
+    else throw new Error("unexpected state for getLineInfo.");
+
+    var output: Filter.Line[] = [];
+    var lowerAdjustment: number;
+    if(reset) lowerAdjustment = 0;
+    else lowerAdjustment = lines.length - previousLines.length;
+
+    // Changes are essentially a partition of the file that is now invalidated
+    // and must be reprocessed. We can reuse the surrounding two partitions,
+    // which will hopefully be the vast majority.
+    var upperPartition: Filter.Line[] = [];
+    if(changes.oldRange.start.row > 0) {
+      upperPartition = previousLines.slice(0, changes.oldRange.start.row);
+    }
+
+    output = output.concat(upperPartition);
+    var newExtent: number = changes.newRange.end.row - changes.newRange.start.row;
+    for(var i = 0; i <= newExtent; i++) {
+      const row = changes.newRange.start.row + i;
+      const currentLine = lines[row];
+
+      const result = parser.parseLine({itemData: itemData, lineText: currentLine,
+          row: row, filePath: this.editor.buffer.getPath()});
+      assert(result, "bad times.");
+      output.push(result);
+    }
+
+    var lowerPartition: Filter.Line[];
+    if(reset) {
+      lowerPartition = [];
+    } else {
+      const remaining = lines.length - output.length;
+      lowerPartition = previousLines.splice(previousLines.length - remaining,
+          previousLines.length);
+    }
+
+    const delta = new Point(lowerAdjustment, 0);
+    lowerPartition.forEach((line) => {
+      this.translateLineRanges(line, delta);
+      output.push(line);
+    });
+
+    emitter.emit<Filter.Params.DataUpdate>("poe-did-process-filter",
+        { editorID: this.editor.buffer.id, lines: output });
+    return output;
   }
 }